requirements.txt
azure-cosmos
azure-identity
azure-keyvault-secrets